requirements.txt
Django==3.2.15
djangorestframework==3.14.0
dj-rest-auth==2.2.5
django-allauth==0.51.0
selenium==4.5.0
